Move SearchMethods To class
