+++ iTunesStore/Controllers/LandscapeViewController.swift
@@ -44,7 +44,7 @@ class LandscapeViewController: UIViewController {
     @IBOutlet weak var scrollView: UIScrollView!
     @IBOutlet weak var pageControl: UIPageControl!
     
-    var searchResults = [SearchResult]()
+    var search: Search!
     private var firstTime = true
     private var downloads = [URLSessionDownloadTask]() // все активные объекты URLSessionDownloadTask
     
@@ -68,7 +68,7 @@ class LandscapeViewController: UIViewController {
         )
         if firstTime {
             firstTime = false
-            tileButtons(searchResults)
+            tileButtons(search.searchResults)
         }
     }
